Feature/auth

+++ src/core/use-cases/removeList.ts
@@ -1,8 +1,13 @@
 import { CoreError } from 'core/errors/CoreError';
 import { ListRepository } from 'core/interfaces/repositories/ListRepository';
+import { UserEntity } from '../entities/User';
 
-export function buildRemoveList({ listRepository }: Dependencies) {
-  return async ({ listId }: Input): Promise<void> => {
+export function removeListUsecaseFactory({ listRepository }: Dependencies) {
+  return async ({ listId, creator }: Input): Promise<void> => {
+    if (!creator) {
+      throw new CoreError('Forbidden');
+    }
+
     if (!listId) {
       throw new CoreError('listId is required.');
     }
@@ -13,7 +18,11 @@ export function buildRemoveList({ listRepository }: Dependencies) {
       throw new CoreError(`No such list found. (${listId})`);
     }
 
-    await listRepository.removeList(listId);
+    if (creator?.id !== list.creator.id) {
+      throw new CoreError('Forbidden');
+    }
+
+    await listRepository.removeList(listId, { creator });
   };
 }
 
@@ -22,4 +31,5 @@ type Dependencies = {
 };
 type Input = {
   listId?: string;
+  creator?: UserEntity | null;
 };
